+++ infra/cdk/stacks/storage_stack.py
@@ -76,7 +76,21 @@ class ClawsStorageStack(cdk.Stack):
             time_to_live_attribute="ttl",
         )
 
+        # DynamoDB table for scheduled watches (TTL-enabled)
+        self.watches_table = dynamodb.Table(
+            self, "WatchesTable",
+            table_name="claws-watches",
+            partition_key=dynamodb.Attribute(
+                name="watch_id",
+                type=dynamodb.AttributeType.STRING,
+            ),
+            billing_mode=dynamodb.BillingMode.PAY_PER_REQUEST,
+            removal_policy=RemovalPolicy.DESTROY,
+            time_to_live_attribute="ttl",
+        )
+
         # Outputs
         cdk.CfnOutput(self, "RunsBucketName", value=self.runs_bucket.bucket_name)
         cdk.CfnOutput(self, "PlansTableName", value=self.plans_table.table_name)
         cdk.CfnOutput(self, "SchemasTableName", value=self.schemas_table.table_name)
+        cdk.CfnOutput(self, "WatchesTableName", value=self.watches_table.table_name)